+++ test/obb/obb_tree_tool.cpp
@@ -60,10 +60,10 @@ static void usage( bool err = true )
 }
 
 #if defined(_MSC_VER)  || defined(__MINGW32__)
-static void memory_use( unsigned long& vsize, unsigned long& rss )
+static void memory_use( unsigned long long& vsize, unsigned long long& rss )
   { vsize = rss = 0; }
 #else
-static void memory_use( unsigned long& vsize, unsigned long& rss )
+static void memory_use( unsigned long long& vsize, unsigned long long& rss )
 {
   char buffer[512];
   int filp = open( "/proc/self/stat", O_RDONLY );
@@ -80,7 +80,7 @@ static void memory_use( unsigned long& vsize, unsigned long& rss )
                   "%*u %*u %*d %*d " // utime stime cutime cstime
                   "%*d %*d %*d "      // priority nice (unused)
                   "%*d %*u "           // itrealval starttime
-                  "%lu %lu",             &vsize, &rss );
+                  "%llu %llu",             &vsize, &rss );
   rss *= getpagesize();
 }
 #endif
@@ -374,7 +374,7 @@ void print_stats( Interface* interface )
   interface->get_entities_by_type( 0, MBVERTEX, verts );
   triangles.merge( verts );
   tree_sets.insert( root );
-  unsigned long set_used, set_amortized, set_store_used, set_store_amortized,
+  unsigned long long set_used, set_amortized, set_store_used, set_store_amortized,
                 set_tag_used, set_tag_amortized, tri_used, tri_amortized;
   interface->estimated_memory_use( tree_sets, 
                                    &set_used, &set_amortized, 
@@ -388,7 +388,7 @@ void print_stats( Interface* interface )
   
   tool.stats( root, std::cout );
   
-  unsigned long real_rss, real_vsize;
+  unsigned long long real_rss, real_vsize;
   memory_use( real_vsize, real_rss );
   
   printf("------------------------------------------------------------------\n");
